/hello page links updated to localhost:8888
The links on the /hello page were pointing to radius, but it is unlikely
anyone will be running this code on radius any time soon. More likely,
this code would run on your laptop, aka "localhost". This patch fixes
the links to point to localhost.

For HTML pages, relative links are possible, so the link could just go
to, for example, "/status", and the browser would fill in the
appropriate hostname and port number.